@@ -18,7 +18,7 @@
 		"class": "BalBuf\\ComposerWP\\Plugin"
 	},
 	"require": {
-		"php": "^5.3",
+		"php": "^5.4",
 		"composer-plugin-api": "^1.0"
 	}
 }
