@@ -9,7 +9,7 @@ namespace BalBuf\ComposerWP\Repository\Config;
 class ZipRepositoryConfig extends RepositoryConfig {
 
 	const repoType = 'wp-zip';
-	protected static $configDefaults = array(
+	protected static $configDefaults = [
 		// rel or absolute path
 		'url' => null,
 		// ssh [user@]hostname
@@ -17,11 +17,11 @@ class ZipRepositoryConfig extends RepositoryConfig {
 		// types and vendors that will be recognized
 		// {type} => {vendors}
 		// types must be any of the keys below - vendors can be anything
-		'package-types' => array(
+		'package-types' => [
 			'wordpress-plugin' => 'wordpress-plugin',
 			'wordpress-muplugin' => 'wordpress-muplugin',
 			'wordpress-theme' => 'wordpress-theme',
-		),
+		],
 		// max depth to traverse directories looking for zip files
 		// falsey or negative value for no max depth
 		'max-depth' => null,
@@ -31,6 +31,6 @@ class ZipRepositoryConfig extends RepositoryConfig {
 		'version-filter' => null,
 		// callable that will receive the package object to manipulate
 		'package-filter' => null,
-	);
+	];
 
 }
